backend/requirements.txt
# 抖音直播间评论监控系统 - 前端

## 项目简介

本项目是一个基于Vue 3和Element Plus的抖音直播间评论监控系统前端应用，可实现直播间评论的实时监控、情感分析和语音回复功能。

## 技术栈

- Vue 3
- Vite
- Element Plus
- Pinia
- Socket.IO
- Tailwind CSS
- Axios

## 项目结构
src/
├── api/                 # API请求
├── assets/              # 静态资源
├── components/          # 组件
├── composables/         # 组合式函数
├── config/              # 配置
├── plugins/             # 插件
├── router/              # 路由
├── stores/              # 状态管理
├── utils/               # 工具函数
├── views/               # 视图
├── App.vue              # 根组件
└── main.js              # 入口文件
## 安装与运行

1. 安装依赖npm install
2. 开发环境运行npm run dev
3. 生产环境构建npm run build
4. 预览构建结果npm run preview
## 配置

在`.env.development`和`.env.production`中配置后端API和Socket.IO地址：
VITE_API_BASE_URL=http://localhost:5000
VITE_SOCKET_URL=http://localhost:5000    Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.6
redis==4.6.0
celery==5.3.1
PyJWT==2.8.0
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==20.1.0
pytest==7.4.0    